Add blanket implementation of Hashable for incrementalmerkletree::Hashable + HashSer
