+++ zcash_primitives/src/merkle_tree.rs
@@ -1,7 +1,7 @@
 //! Implementation of a Merkle tree of commitments used to prove the existence of notes.
 
 use byteorder::{LittleEndian, ReadBytesExt};
-use incrementalmerkletree::bridgetree;
+use incrementalmerkletree::{self, bridgetree, Altitude};
 use std::collections::VecDeque;
 use std::io::{self, Read, Write};
 
@@ -28,6 +28,45 @@ pub trait Hashable: Clone + Copy {
     /// Returns the empty root for the given depth.
     fn empty_root(_: usize) -> Self;
 }
+
+/// A hashable node within a Merkle tree.
+pub trait HashSer {
+    /// Parses a node from the given byte source.
+    fn read<R: Read>(reader: R) -> io::Result<Self>
+    where
+        Self: Sized;
+
+    /// Serializes this node.
+    fn write<W: Write>(&self, writer: W) -> io::Result<()>;
+}
+
+impl<T> Hashable for T where T: incrementalmerkletree::Hashable + HashSer + Copy {
+    /// Parses a node from the given byte source.
+    fn read<R: Read>(reader: R) -> io::Result<Self> {
+        <Self as HashSer>::read(reader)
+    }
+
+    /// Serializes this node.
+    fn write<W: Write>(&self, writer: W) -> io::Result<()> {
+        <Self as HashSer>::write(self, writer)
+    }
+
+    /// Returns the parent node within the tree of the two given nodes.
+    fn combine(alt: usize, lhs: &Self, rhs: &Self) -> Self {
+        <Self as incrementalmerkletree::Hashable>::combine(Altitude::from(alt as u8), lhs, rhs)
+    }
+
+    /// Returns a blank leaf node.
+    fn blank() -> Self {
+        <Self as incrementalmerkletree::Hashable>::empty_leaf()
+    }
+
+    /// Returns the empty root for the given depth.
+    fn empty_root(alt: usize) -> Self {
+        <Self as incrementalmerkletree::Hashable>::empty_root(Altitude::from(alt as u8))
+    }
+}
+
 
 struct PathFiller<Node: Hashable> {
     queue: VecDeque<Node>,
